+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-13T13:49:16.267Z")
+    created: Optional[datetime] = parse_datetime("2025-08-14T05:22:42.646Z")
     r"""The date and time the resource was created"""