+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2HV2JCM2W03WZMPQTMJW47M"
+        "01K2KGFR52T1R3FEG68Y35PRR5"
     )
     r"""The id of the resource"""