+++ packages/orq-rc/RELEASES.md
@@ -2279,3 +2279,13 @@ Based on:
 - [python v3.11.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v3.11.0-rc.10] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.10 - packages/orq-rc
+
+## 2025-08-14 05:24:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.603.2 (2.681.8) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.11.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v3.11.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.11 - packages/orq-rc